Feature/show held components (#6)
Setting to toggle an additional column in the crafting table which will display which actors are holding relevant components for each recipe.
Setting to toggle support for the Party-Inventory module (https://github.com/teroparvinen/foundry-party-inventory).
Toggle to filter the crafting table, only displaying recipes which the party have the required components.
Known issue - Items created in party-inventory and moved across to an actor do not appear on the actors sheet but are still found by this module and displayed in the crafting table. Further testing is needed.